[so/testing] Reorganize test files
